requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.1.4
django-cors-headers==3.13.0
django-environ==0.9.0
django-redis==5.2.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
exceptiongroup==1.1.0
iniconfig==1.1.1
kombu==5.2.4
packaging==22.0
pluggy==1.0.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
pytest==7.2.0
pytest-sugar==0.9.6
pytz==2022.7
redis==4.4.0
six==1.16.0
sqlparse==0.4.3
termcolor==2.1.1
tomli==2.0.1
uwsgi
vine==5.0.0
wcwidth==0.2.5
